headers: extend badges for devel_job info
Loading the newly introduced devel_job information by the buildfarm
requires an extension to the data structure of badges.
The current implementation looks for the fields "aggregated_status",
"history" and "results_latest_build" to fill the badge data.
This only happens if the devel_job information was successfully
appended to the manifest in the previous step.